+++ resources/js/App.vue
@@ -1,5 +1,5 @@
 <script setup>
-import Counter from './components/Counter.vue';
+import MyCounter from './components/MyCounter.vue';
 import logo from '../images/logo-with-shadow.png';
 </script>
 
@@ -7,7 +7,7 @@ import logo from '../images/logo-with-shadow.png';
   <h1>Hello World</h1>
   <img :src="logo" />
   <img src="../images/logo-with-shadow.png" />
-  <Counter />
+  <MyCounter />
   <p class="text-primary">Hello</p>
   <button type="button" class="btn btn-primary">ボタン</button>
   <button type="button" class="btn btn-primary rounded-pill">長丸ボタン</button>
