+++ resources/views/welcome.blade.php
